+++ src/wtforms/__init__.py
@@ -36,7 +36,7 @@ from wtforms.fields.simple import URLField
 from wtforms.form import Form
 from wtforms.validators import ValidationError
 
-__version__ = "3.2.0"
+__version__ = "3.2.1"
 
 __all__ = [
     "validators",