+++ src/wtforms/fields/choices.py
@@ -114,11 +114,11 @@ class SelectField(SelectFieldBase):
         if not choices:
             _choices = []
 
-        elif isinstance(choices[0], (list, tuple)):
+        elif isinstance(choices[0], list | tuple):
             _choices = choices
 
         else:
-            _choices = zip(choices, choices)
+            _choices = zip(choices, choices, strict=False)
 
         for value, label, *other_args in _choices:
             selected = self.coerce(value) == self.data
@@ -168,11 +168,11 @@ class SelectMultipleField(SelectField):
         if not choices:
             _choices = []
 
-        elif isinstance(choices[0], (list, tuple)):
+        elif isinstance(choices[0], list | tuple):
             _choices = choices
 
         else:
-            _choices = zip(choices, choices)
+            _choices = zip(choices, choices, strict=False)
 
         for value, label, *other_args in _choices:
             selected = self.data is not None and self.coerce(value) in self.data