+++ CHANGES.rst
@@ -6,6 +6,7 @@ Unreleased
 Unreleased
 
 - Fix :class:`~validators.Disabled` validation with provided formdata. :pr:`880`
+- End support for Python 3.9, start support for Python 3.14. :pr:`883`
 
 Version 3.2.1
 -------------