+++ CHANGES.rst
@@ -7,6 +7,15 @@ Unreleased
 
 - Add :class:`~validators.DateRange`. :pr:`787`
 
+Version 3.2.2
+-------------
+
+Released 2026-05-03
+
+- Fix :class:`~validators.Disabled` validation with provided formdata. :pr:`880`
+- End support for Python 3.9, start support for Python 3.14. :pr:`883`
+- Add Tamil and Serbian translations.
+
 Version 3.2.1
 -------------